Revert unwanted changes and update documentation
