+++ base/src/main/java/io/spine/validate/AlternativeFieldValidator.java
@@ -139,10 +139,9 @@ final class AlternativeFieldValidator implements Logging {
             violations.add(notFound);
             return false;
         }
-        FieldValueChange change = FieldValueChange.withoutPreviousValue(fieldValue.get());
-        FieldValidator<?> fieldValidator = change.createValidatorAssumingRequired();
+        FieldValidator<?> fieldValidator = fieldValue.get()
+                                                     .createValidatorAssumingRequired();
         List<ConstraintViolation> violations = fieldValidator.validate();
-
         // Do not add violations to the results because we have options.
         // The violation would be that none of the field or combinations is defined.
