+++ base/src/main/java/io/spine/validate/FieldValidator.java
@@ -46,7 +46,7 @@ abstract class FieldValidator<V> implements Logging {
     private static final String ENTITY_ID_REPEATED_FIELD_MSG =
             "Entity ID must not be a repeated field.";
 
-    private final FieldValueChange fieldValueChange;
+    private final FieldValue value;
     private final FieldDeclaration declaration;
     private final ImmutableList<V> values;
 
@@ -74,7 +74,7 @@ abstract class FieldValidator<V> implements Logging {
      *         if this constraint is not set explicitly
      */
     protected FieldValidator(FieldValue fieldValue, boolean assumeRequired) {
-        this.fieldValueChange = FieldValueChange.withoutPreviousValue(fieldValue);
+        this.value = fieldValue;
         this.declaration = fieldValue.declaration();
         this.values = fieldValue.asList();
         this.assumeRequired = assumeRequired;
@@ -85,19 +85,6 @@ abstract class FieldValidator<V> implements Logging {
         this.setOnce = fieldValue.valueOf(OptionsProto.setOnce);
     }
 
-    protected FieldValidator(FieldValueChange valueChange, boolean assumeRequired) {
-        this.fieldValueChange = valueChange;
-        FieldValue desiredValue = valueChange.newValue();
-        this.declaration = desiredValue.declaration();
-        this.values = desiredValue.asList();
-        this.assumeRequired = assumeRequired;
-        this.required = desiredValue.valueOf(OptionsProto.required);
-        this.ifMissingOption = desiredValue.valueOf(OptionsProto.ifMissing);
-        this.validate = desiredValue.valueOf(OptionsProto.valid);
-        this.ifInvalid = desiredValue.valueOf(OptionsProto.ifInvalid);
-        this.setOnce = desiredValue.valueOf(OptionsProto.setOnce);
-    }
-
     /**
      * Checks if the value of the validated field is not set.
      *
@@ -105,7 +92,7 @@ abstract class FieldValidator<V> implements Logging {
      *
      * @return {@code true} if the field value is not set and {@code false} otherwise
      */
-    boolean fieldValueNotSet(FieldValue value) {
+    boolean fieldValueNotSet() {
         List<V> values = value.asList();
         boolean valueNotSet =
                 values.isEmpty()
@@ -144,7 +131,6 @@ abstract class FieldValidator<V> implements Logging {
      */
     protected final List<ConstraintViolation> validate() {
         checkIfRequiredAndNotSet();
-        validateSetOnce();
         if (isRequiredId()) {
             validateEntityId();
         }
@@ -153,27 +139,6 @@ abstract class FieldValidator<V> implements Logging {
         }
         List<ConstraintViolation> result = assembleViolations();
         return result;
-    }
-
-    private void validateSetOnce() {
-        FieldValue previousValue = fieldValueChange.previousValue();
-        FieldValue desiredValue = fieldValueChange.newValue();
-        boolean previousValueSet = !fieldValueNotSet(previousValue);
-        boolean desiredValueSet = !fieldValueNotSet(desiredValue);
-        boolean valueChanged = previousValueSet && desiredValueSet;
-        if(setOnce && valueChanged){
-            String fieldName = previousValue.declaration()
-                                            .name()
-                                            .javaCase();
-            ConstraintViolation setOnceViolation = ConstraintViolation
-                    .newBuilder()
-                    .setMsgFormat("%s has (set_once) = true, can't change its value from %s to %s")
-                    .addParam(fieldName)
-                    .addParam(previousValue.toString())
-                    .addParam(desiredValue.toString())
-                    .build();
-            addViolation(setOnceViolation);
-        }
     }
 
     /**
@@ -206,7 +171,7 @@ abstract class FieldValidator<V> implements Logging {
             addViolation(violation);
             return;
         }
-        if (fieldValueNotSet(fieldValueChange.newValue())) {
+        if (fieldValueNotSet()) {
             addViolation(newViolation(ifMissingOption));
         }
     }
@@ -242,7 +207,7 @@ abstract class FieldValidator<V> implements Logging {
             }
             return;
         }
-        if (fieldValueNotSet(fieldValueChange.newValue())) {
+        if (fieldValueNotSet()) {
             addViolation(newViolation(ifMissingOption));
         }
     }
@@ -319,9 +284,8 @@ abstract class FieldValidator<V> implements Logging {
      * @return {@code true} if the field is a required entity ID, {@code false} otherwise
      */
     private boolean isRequiredEntityId() {
-        boolean requiredSetExplicitly = fieldValueChange.newValue()
-                                                        .option(OptionsProto.required)
-                                                        .isExplicitlySet();
+        boolean requiredSetExplicitly = value.option(OptionsProto.required)
+                                             .isExplicitlySet();
         boolean notRequired = !required && requiredSetExplicitly;
         return declaration.isEntityId() && !notRequired;
     }
@@ -332,7 +296,7 @@ abstract class FieldValidator<V> implements Logging {
      * @return the field context
      */
     protected FieldContext getFieldContext() {
-        return fieldValueChange.newValue().context();
+        return value.context();
     }
 
     /** Returns a path to the current field. */
@@ -343,9 +307,5 @@ abstract class FieldValidator<V> implements Logging {
     /** Returns the declaration of the validated field. */
     protected FieldDeclaration field() {
         return declaration;
-    }
-
-    final FieldValue desiredValue(){
-        return fieldValueChange.newValue();
     }
 }
