+++ base/src/main/java/io/spine/validate/MessageFieldValidator.java
@@ -60,20 +60,6 @@ class MessageFieldValidator extends FieldValidator<Message> {
         this.timeConstraint = fieldValue.valueOf(OptionsProto.when);
     }
 
-    /**
-     * Creates a new validator instance.
-     *
-     * @param fieldValueChange
-     *         the change of the field value to validate
-     * @param assumeRequired
-     *         if {@code true} the validator would assume that the field is required even if
-     *         such constraint is not explicitly set
-     */
-    MessageFieldValidator(FieldValueChange fieldValueChange, boolean assumeRequired) {
-        super(fieldValueChange, assumeRequired);
-        this.timeConstraint = fieldValueChange.newValue().valueOf(OptionsProto.when);
-    }
-
     @Override
     protected void validateOwnRules() {
         boolean validateFields = shouldValidateFields();
@@ -85,7 +71,7 @@ class MessageFieldValidator extends FieldValidator<Message> {
     }
 
     private boolean shouldValidateFields() {
-        return getValidateOption() && !fieldValueNotSet(desiredValue());
+        return getValidateOption() && !fieldValueNotSet();
     }
 
     @Override
